Restructured updater to be more maintainable
