+++ src/main/lib/updater/index.js
@@ -1,0 +1,10 @@
+/**
+ * @file Manages exports from updater module
+ */
+import { start } from "./updater";
+import Store from "./dataStore";
+
+export default {
+  "preinit": start,
+  "Provider": Store.Provider,
+};
